requirements.txt
duckdb
pandas
numpy
matplotlib
seaborn
psycopg2
sqlalchemy
streamlit
scikit-learn
geopandas
folium
geoalchemy2
shapely
xgboost
python-dotenv
altair
plotly